Refactor project structure and improve report layout
