+++ README.md
@@ -3,8 +3,9 @@
 This project demonstrates how to collect system facts from hosts using Ansible
 and render the results in a small React application. The playbook gathers the
 list of system users and listening network ports, then saves the collected data
-as JSON. A Python helper script runs the playbook and generates a React based
-web page that displays the results.
+as JSON. A Python helper script located in the `scripts/` directory runs the
+playbook and generates a React-based web page using templates stored in
+`templates/`.
 
 ## Requirements
 - Python 3
@@ -25,7 +26,7 @@ site in the `results` directory and launch a local nginx server to serve it
 remotely:
 
 ```bash
-python3 collect_and_visualize.py
+python3 scripts/collect_and_visualize.py
 ```
 
 After the script completes, visit `http://localhost:8080` to view the collected
