+++ templates/index.html.j2
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset='utf-8'>
+    <title>System Facts</title>
+    <script crossorigin src="https://unpkg.com/react@17/umd/react.production.min.js"></script>
+    <script crossorigin src="https://unpkg.com/react-dom@17/umd/react-dom.production.min.js"></script>
+    <script crossorigin src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
+    <style>
+        body { font-family: Arial, sans-serif; }
+        table { border-collapse: collapse; width: 100%; margin-top: 1em; }
+        th, td { border: 1px solid #ddd; padding: 8px; }
+        th { background-color: #f2f2f2; text-align: left; }
+        ul { list-style-type: none; padding: 0; margin: 0; }
+        li { margin: 0.2em 0; }
+    </style>
+</head>
+<body>
+    <div id="root"></div>
+    <script type="text/babel">
+    fetch('data.json')
+      .then(response => response.json())
+      .then(data => {
+        function App() {
+          return (
+            <div>
+              <h1>Collected System Facts</h1>
+              <table>
+                <thead>
+                  <tr>
+                    <th>Hostname</th>
+                    <th>Users</th>
+                    <th>Listening Ports</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {data.map((h, i) => (
+                    <tr key={i}>
+                      <td>{h.hostname}</td>
+                      <td>
+                        <ul>{h.users.map((u, j) => <li key={j}>{u}</li>)}</ul>
+                      </td>
+                      <td>
+                        <ul>{h.ports.map((p, j) => <li key={j}>{p}</li>)}</ul>
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          );
+        }
+        ReactDOM.render(<App />, document.getElementById('root'));
+      });
+    </script>
+</body>
+</html>
